% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional-estimators.R
\name{bestParameters}
\alias{bestParameters}
\title{bestParameters}
\usage{
bestParameters(y, nbK = 10, type = "MAD", sdEta = TRUE)
}
\arguments{
\item{y}{A time-series obtained by the dataRWAR function}

\item{nbK}{number of diff k elements to consider}

\item{type}{type of robust variance estimator (MAD, S or Q)}

\item{sdEta}{if sdEta = FALSE there is no random walk}
}
\value{
a list with an estimation of the best parameters for Eta2, Nu2 and phi
}
\description{
iteration of the least square criterion for a grid of the phi parameter
}
\examples{
bestParameters(dataRWAR(10000, sdEta = 0.2, sdNu = 0.1, phi = 0.3,
type = "rand1", nbSeg = 10)$y)
}
