% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional-estimators.R
\name{cost}
\alias{cost}
\title{L2 error estimation}
\usage{
cost(v, sdEta, sdNu, phi)
}
\arguments{
\item{v}{the estimated variances of the diff k operator}

\item{sdEta}{standard deviation in Random Walk}

\item{sdNu}{standard deviation in AR(1)}

\item{phi}{the autocorrelative AR(1) parameter}
}
\value{
the value of the sum of squares
}
\description{
the least-square value
}
