% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRWAR.R
\name{dataRWAR}
\alias{dataRWAR}
\title{Title}
\usage{
dataRWAR(
  n = 1000,
  poisParam = 0.01,
  meanGap = 10,
  phi = 0.98,
  sdEta = 1,
  sdNu = 1
)
}
\arguments{
\item{n}{}

\item{poisParam}{}

\item{meanGap}{}

\item{phi}{}

\item{sdEta}{}

\item{sdNu}{}
}
\value{

}
\description{
Title
}
