% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateParameters.R
\name{estimateParameters}
\alias{estimateParameters}
\title{Estimate parameter in the Random Walk Autoregressive model}
\usage{
estimateParameters(y, K = 20)
}
\arguments{
\item{y}{A vector of observations}

\item{K}{The number of lags to run the estimation over. Default set at 20.}
}
\value{
A list containing $sigmaEta, the sd of the Random Walk Component, $sigmaNu, the sd of the AR noise, $phi, the autocorrelation parameter.
}
\description{
This function perform robust estimation of parameters in the Random Walk plus Autoregressive model using
a method of moments estimator.
}
\examples{
set.seed(42)
y <- dataRWAR(n = 1e4, poisParam = .01, phi = .7, sdEta = 4, sdNu = 3)$y
estimateParameters(y)
}
