% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.DeCAFSout}
\alias{plot.DeCAFSout}
\title{DeCAFS Plotting}
\usage{
\method{plot}{DeCAFSout}(x, ...)
}
\arguments{
\item{x}{the output object from a DeCAFS call}

\item{...}{Additional graphical parameters to be passed down to the plot function}
}
\value{
An R plot
}
\description{
DeCAFS output plotting method.
}
\examples{
set.seed(42)
Y <- dataRWAR(n = 1e3, phi = .5, sdEta = 1, sdNu = 3,  jumpSize = 15, type = "updown", nbSeg = 5)
res = DeCAFS(Y$y)
plot(res, type = "l")

}
