% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarioGenerator.R
\name{scenarioGenerator}
\alias{scenarioGenerator}
\title{Generate a piecewise constant signal of a given length}
\usage{
scenarioGenerator(
  n,
  type = c("none", "up", "updown", "rand1"),
  nbSeg = 20,
  jumpSize = 1
)
}
\arguments{
\item{n}{The length of the sequence of observations.}

\item{type}{Possible change scenarios for the jump structure}

\item{nbSeg}{Number of segments}

\item{jumpSize}{Maximum magnitude of a change}
}
\value{
a sequence of N values for the piecewise constant signal
}
\description{
Generate a piecewise constant signal of a given length
}
\examples{
scenarioGenerator(1e3, "rand1")

}
