% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRWAR.R
\name{dataRWAR}
\alias{dataRWAR}
\title{Generate a Random Walk + AR realization}
\usage{
dataRWAR(
  n = 1000,
  poisParam = 0.01,
  meanGap = 10,
  phi = 0.98,
  sdEta = 1,
  sdNu = 1
)
}
\arguments{
\item{n}{The length of the sequence of observations.}

\item{poisParam}{A poisson parameter regulating the probability of observing a change.}

\item{meanGap}{The average magnitude of a change.}

\item{phi}{The autocorrelation parameter \eqn{\phi}}

\item{sdEta}{The standard deviation of the Random Walk Component}

\item{sdNu}{The standard deviation of the Autocorrelated noise}
}
\value{

}
\description{
Generate a Realization of the RWAR model:
\deqn{y_t = \mu_t + \epsilon_t}
where 
\deqn{\mu_t = \mu_{t-1} + \eta_t + \delta_t, \quad \eta_t \sim N(0, \sigma_\eta^2), \ \delta_t \ \in R}
and
\deqn{\epsilon_t = \phi \epsilon_{t-1} + \nu_t \quad \nu_t \sim N(0, \sigma_\nu^2)}
}
\examples{
library(ggplot2)
set.seed(42)
Y = dataRWAR(n = 1e3, poisParam = .01, meanGap = 15, phi = .5, sdEta = 3, sdNu = 1)
y = Y$y
ggplot(data.frame(t = 1:length(y), y), aes(x = t, y = y)) +
  geom_point() +
  geom_vline(xintercept = Y$changepoints, col = 4,  lty = 3)
}
