% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSinusoidal.R
\name{dataSinusoidal}
\alias{dataSinusoidal}
\title{Generating data from a sinusoidal model with changes}
\usage{
dataSinusoidal(
  n,
  poisParam = 0.01,
  meanGap = 10,
  amplitude = 1,
  frequency = 1,
  phase = 0,
  sd = 1
)
}
\arguments{
\item{n}{The length of the sequence of observations.}

\item{poisParam}{A poisson parameter regulating the probability of observing a change.}

\item{meanGap}{The average magnitude of a change.}

\item{amplitude}{The amplitude of the sinusoid}

\item{frequency}{The angular frequency of the sinusoid}

\item{phase}{where the signal starts at time t = 0}

\item{sd}{standard deviation of the noise added on top of the signal}
}
\value{
The sequence of observation, the sinusoid signal at the basis, the changepoints location.
}
\description{
This function generates a sequence of observation from a sinusoidal model with changes. This can be used as an example 
for model misspecification.
}
\examples{
Y <- dataSinusoidal(1e4, poisParam = .0005, meanGap = 5, frequency = 2 * pi / 1e3, amplitude = 10, sd = 2)
res <- DeCAFS(Y$y)
par(mfrow = c(2, 1))
plot(Y$y, col = "grey")
lines(res$signal, col = "blue", lwd = 2)
lines(Y$signal, col = "red", lwd = 2, lty = 2)
abline(v = res$changepoints, col = 4)
abline(v = Y$changepoints, col = 2, lty = 2)
plot(Y$y[-1] - Y$y[-1e4])
abline(v = Y$changepoints, col = 2, lty = 2)
par(mfrow = c(1, 1))
}
